Moved columnwidths list to container base.
