--- src/Pdf/Containers/Tables/README.md
@@ -1,11 +1,0 @@
-# Tables
-
-## Minimum Functionality Goals
-- [ ] Multiple page support
-- [ ] Option to continue headers onto the start of every new page.
-- [ ] Auto column widths
-- [x] Consistent row heights
-- [x] Percentage-based row widths
-- [x] Row Separators
-- [ ] Striped rows
-- [ ] All bordered cell rows
