+++ src/Pdf/Views/ITable.cs
@@ -8,14 +8,14 @@ public interface ITable : IContainer
     ITable WithColumnWidths(Action<ColumnBuilder> configureColumns);
 
     /// <summary>
-    ///
+    /// Adds a header to the vStack
     /// </summary>
     /// <param name="configure"></param>
     /// <returns></returns>
     ITable WithHeader(Action<VStack> configure);
 
     /// <summary>
-    ///
+    /// Adds a header to the vStack with custom styling
     /// </summary>
     /// <param name="configure"></param>
     /// <param name="configureStyle"></param>
@@ -23,22 +23,17 @@ public interface ITable : IContainer
     ITable WithHeader(Action<VStack> configure, Func<BlockStyle, BlockStyle> configureStyle);
 
     /// <summary>
-    ///
+    /// Adds a footer to the vStack
     /// </summary>
     /// <param name="configure"></param>
     /// <returns></returns>
     public ITable WithFooter(Action<VStack> configure);
 
     /// <summary>
-    ///
+    /// Adds a footer to the vStack with custom styling
     /// </summary>
     /// <param name="configure"></param>
     /// <param name="configureStyle"></param>
     /// <returns></returns>
     public ITable WithFooter(Action<VStack> configure, Func<BlockStyle, BlockStyle> configureStyle);
-}
-
-public interface IRow
-{
-    IRow WithColumnWidths(Action<ColumnBuilder> configureColumns);
 }
